refactor: Pentagram Node 컴포넌트 분리 - 위치 조정 래퍼 컴포넌트, 클릭 이벤트 전달 컴포넌트
